Move configuration into its own package
